% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_na.R
\name{explore_na}
\alias{explore_na}
\title{Explore missing values in a dataset}
\usage{
explore_na(
  df,
  return = "plot",
  probabilities = TRUE,
  keep_all = FALSE,
  colour_thresholds = c(0.1, 0.3),
  zoom_in = FALSE,
  reference_lines = TRUE
)
}
\arguments{
\item{df}{dataset, of the class data.frame}

\item{return}{Either "plot" or "table", depending what you want the function to return}

\item{probabilities}{Do you want the amount missing reported in absolute (nrow) or relative (percentage)terms}

\item{keep_all}{Whether to include the columns which are not missing anything ie n_missing = 0}

\item{colour_thresholds}{The visuals give a RAG rating for the missingness depending on these thresholds}

\item{zoom_in}{Should the amount missing axis max out at 100 percent or the greatest non-zero amount missing?}

\item{reference_lines}{Should reference lines be drawn on the visual to show the thresholds used for the colours}
}
\value{
A ggplot object
}
\description{
Primarily a visual tool, but also can tabulate the same, which allows exploration of the amount of missingness in a dataset. The visual shows the amount of missing data in each column.
}
\examples{
my_data <- data.frame(
  x0 = 1:11
  , x1 = c(NA, 1:10)
  , x2 = c(rep(NA, 2), 1:9)
  , x3 = c(rep(NA, 3), 1:8)
  , x4 = c(rep(NA, 4), 1:7)
)
my_data |> explore_na()
my_data |> explore_na(keep_all = TRUE)
my_data |> explore_na(zoom_in = TRUE)

}
