% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorebunny-package.R
\docType{package}
\name{explorebunny-package}
\alias{explorebunny}
\alias{explorebunny-package}
\title{explorebunny: Exploratory Data Analysis (Yet another EDA package eh)}
\description{
Explore datasets (just how I like it).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danielbraddock.github.io/explorebunny/}
}

}
\author{
\strong{Maintainer}: Daniel Braddock \email{nobody@noreply.com}

}
\keyword{internal}
