% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{Who am I talking to? Defaults to "world"}
}
\value{
A greeting to "name"
}
\description{
Greets whoever you ask it to greet via name
}
\examples{
hello()
hello("you")
hello("hello")
hello("Mr Sandman")
}
